+++ sdk/keyvault/keyvault-certificates/src/certificatesModels.ts
@@ -813,3 +813,91 @@ export type RequireAtLeastOne<T> = {
  * which are: emailAddress, name or phone.
  */
 export type CertificateContact = RequireAtLeastOne<CertificateContactAll> | undefined;
+
+/** Known values of {@link CertificateKeyCurveName} that the service accepts. */
+export enum KnownCertificateKeyCurveNames {
+  /**
+   * P-256 Key Curve.
+   */
+  P256 = "P-256",
+  /**
+   * P-384 Key Curve.
+   */
+  P384 = "P-384",
+  /**
+   * P-521 Key Curve.
+   */
+  P521 = "P-521",
+  /**
+   * P-256K Key Curve.
+   */
+  P256K = "P-256K"
+}
+
+/** Known values of {@link CertificateKeyType} that the service accepts. */
+export enum KnownCertificateKeyTypes {
+  /**
+   * EC Key Type.
+   */
+  EC = "EC",
+  /**
+   * EC-HSM Key Type.
+   */
+  ECHSM = "EC-HSM",
+  /**
+   * RSA Key Type.
+   */
+  RSA = "RSA",
+  /**
+   * RSA-HSM Key Type.
+   */
+  RSAHSM = "RSA-HSM",
+  /**
+   * oct Key Type
+   */
+  Oct = "oct",
+  /**
+   * oct-HSM Key Type
+   */
+  OctHSM = "oct-HSM"
+}
+
+/** Known values of {@link KeyUsageType} that the service accepts. */
+export const enum KnownKeyUsageTypes {
+  /**
+   * DigitalSignature Usage Type.
+   */
+  DigitalSignature = "digitalSignature",
+  /**
+   * NonRepudiation Usage Type.
+   */
+  NonRepudiation = "nonRepudiation",
+  /**
+   * KeyEncipherment Usage Type.
+   */
+  KeyEncipherment = "keyEncipherment",
+  /**
+   * DataEncipherment Usage Type.
+   */
+  DataEncipherment = "dataEncipherment",
+  /**
+   * KeyAgreement Usage Type.
+   */
+  KeyAgreement = "keyAgreement",
+  /**
+   * KeyCertSign Usage Type.
+   */
+  KeyCertSign = "keyCertSign",
+  /**
+   * CRLSign Usage Type.
+   */
+  CRLSign = "cRLSign",
+  /**
+   * EncipherOnly Usage Type.
+   */
+  EncipherOnly = "encipherOnly",
+  /**
+   * DecipherOnly Usage Type.
+   */
+  DecipherOnly = "decipherOnly"
+}