+++ sdk/storage/storage-file-datalake/test/utils.spec.ts
@@ -11,6 +11,7 @@ import {
 } from "../src/utils/utils.common";
 import { record, Recorder } from "@azure-tools/test-recorder";
 import { recorderEnvSetup } from "./utils";
+import { Context } from "mocha";
 dotenv.config();
 
 describe("Utility Helpers", () => {
@@ -40,7 +41,7 @@ describe("Utility Helpers", () => {
     );
   }
 
-  beforeEach(function() {
+  beforeEach(function(this: Context) {
     recorder = record(this, recorderEnvSetup);
   });