+++ sdk/storage/storage-queue/test/utils/index.browser.ts
@@ -13,9 +13,9 @@ export function getGenericQSU(
   const accountNameEnvVar = `${accountType}ACCOUNT_NAME`;
   const accountSASEnvVar = `${accountType}ACCOUNT_SAS`;
 
-  let accountName: string | undefined;
+  const accountName = (self as any).__env__[accountNameEnvVar];
+
   let accountSAS: string | undefined;
-  accountName = (self as any).__env__[accountNameEnvVar];
   accountSAS = (self as any).__env__[accountSASEnvVar];
 
   if (!accountName || !accountSAS || accountName === "" || accountSAS === "") {
@@ -57,7 +57,6 @@ export async function bodyToString(
     readableStreamBody?: NodeJS.ReadableStream;
     blobBody?: Promise<Blob>;
   },
-  // tslint:disable-next-line:variable-name
   _length?: number
 ): Promise<string> {
   const blob = await response.blobBody!;